Address review comments and refactor score component names to make sense
